fix(commerce): 상품 가격 Long 전환에 맞춰 테스트·온톨로지 참조를 고친다
79c46a57 이 product Money 를 Long 으로 바꾸면서 commerce:app 테스트 5곳(BigDecimal 가격)이 컴파일되지 않았고,
온톨로지의 코드 참조(Money.toWon)가 사라진 심볼을 가리켜 OntologyFilesSpec 이 깨졌다(c8a72405 CI 실패).
참조를 value class Money(val amount: Long) 로 옮기고 manifest revision 을 올린다.

+++ commerce/app/src/test/kotlin/com/kgd/commerce/saga/OrderSagaE2ETest.kt
@@ -148,7 +148,7 @@ class OrderSagaE2ETest(
 
     /** 플랫폼(판매자 1, 배송비 0) 상품 + 창고 1 입고 — order 읽기 모델에 들어올 때까지 기다린다 */
     private fun product(name: String, price: Long, stock: Int, requester: ProductRequester = admin): Long {
-        val id = products.execute(CreateProductUseCase.Command(name = name, price = price.toBigDecimal(), stock = stock), requester).id
+        val id = products.execute(CreateProductUseCase.Command(name = name, price = price.toLong(), stock = stock), requester).id
         receiveStock.execute(ReceiveStockUseCase.Command(productId = id, warehouseId = 1L, qty = stock))
         awaitUntil("product_view $id") {
             orderJdbc.queryForList("SELECT price FROM product_view WHERE product_id = ?", Long::class.java, id).singleOrNull() == price
